[SOIN] Sépare la notion de consignation et de notification
...lorsqu'on traite d'un avis donnée sur une mesure
