+++ back/src/bus/consigneAvisMesureDonneDansJournal.ts
@@ -1,6 +1,5 @@
 import { AdaptateurHorloge } from '../infra/adaptateurHorloge.js';
 import { AdaptateurJournal } from '../infra/adaptateurJournal.js';
-import { MessagerieInstantanee } from '../metier/messagerieInstantanee.js';
 import { AvisMesureDonne } from './evenements/avisMesureDonne.js';
 
 export const consigneRetourAvisMesureDonneDansJournal = ({
@@ -21,21 +20,6 @@ export const consigneRetourAvisMesureDonneDansJournal = ({
       },
       type: 'AVIS_MESURE_DONNE',
       date: adaptateurHorloge.maintenant(),
-    });
-  };
-};
-
-export const consigneCommentaireAvisMesureDonneDansMessagerie = ({
-  messagerieInstantanee,
-}: {
-  messagerieInstantanee: MessagerieInstantanee;
-}) => {
-  return async (evenement: AvisMesureDonne) => {
-    if (evenement.retour !== 'NEGATIF') return;
-    await messagerieInstantanee.notifieUnAvisNegatifSurUneMesure({
-      idMesure: evenement.idMesure,
-      titreMesure: evenement.titreMesure,
-      commentaire: evenement.commentaire,
     });
   };
 };
